requirements.txt
streamlit>=1.31.0
pandas>=2.2.0
plotly>=5.18.0
fpdf2>=2.7.5